requirements.txt
kaggle
pandas
sqlalchemy
psycopg2-binary
python-dotenv